Add filter on annotation test to only check constructor members
Annotation should be present on all arguments that the user can specify in a viash config script.
Extra annotations may be present on private vals to provide extra documentation but surely it is not a requirement.

Add MemberInfo to contain information in structured way instead of a tuple
Add shortName method to get the last section of a fullName