Add functionality name and namespace in local dependency config info
Functionality name and namespace are already present in remote dependencies.
Adding them for local dependencies makes it more consistent and provides needed information for the NextflowScript code.

Also clean up code for converting script paths to strings.
All usages of the path needed to convert it to a string, so do it straight away.