Fix escaping BashWrapper pipe characters being stripped (#422)
* Fix escaping BashWrapper pipe characters being stripped

Fix escaping of the included script where a line starting with a pipe character with optional leading spaces is stripped of the leading spaces and pipe character.

Add a function that escapes pipe characters (with leading spaces) and use where stripMargin is called on the script.

* Fix possible future issue with a leading pipe character

Strip off a leading pipe character. It'd cause an extra unwanted pipe.

* Simplify pipe escaping. Always add a pipe after a newline.

Make the code simpler by always adding pipes, whether needed or not.