Add [+-].inf and .nan as valid option for double strings, add the string values as valid enum values for a double field
Add a custom "Double" type of standard number or an enum value
Add +.inf, -.inf, .nan as valid strings when decoding a Double value with circe

This still doesn't handle a yaml with `min: +.inf`, for that we need to pre-parse the yaml before circe even sees it. This can't be handled with a custom decoder.