Feature/sys env wrapper (#457)
* Add wrapper around sys.env and provide access to specific variables

This allows us to track which ones are used and document them properly.
This is also used for the testbenches to set an environment variable during testing.
Setting of variables is blocked during regular code execution.

* Update CHANGELOG.md

* Add a very short testbench

* Remove VIASH_VERSION override after running the MainRunVersionSwitch testbench

* Move SysEnv test code to MainRunVersionSwitch

This prevents concurrency issues

* Add documentation for SysEnv

Add 'nameOverride' annotation that can be used to override field names.
So far it can't set class names. This should probably be added.