Copy static reference pages from the website to here
This way we can track changes in e.g. config mods locally

Odd one out is the cli/index page as it auto generates using a template.
The root of the reference is statically defined in the website repo too. I'm not sure it makes sense to move these too.

Last consideration, it would be nice if the `order` field in the header could be filled in by the website but it seems more complex than that it's worth.