Fix testbench failing only when running locally and while running sbt test
When running only that test locally it was successfully too.
Turns out to be a collision in the docker image name.

Rewritten the test to use the ConfigDeriver functionality.
Lots of the custom code for this test was separately implemented in ConfigDeriver.

Piggy back a minor improvement in the config_mods AppendTest testbench.