Don't output `internal functionality` fields during `config view`
Same goes for `ns list`

Add a custom `deriveConfiguredEncoderStrict` which checks annotations and remove any field annotated with `@internalFunctionality`.

Add a `hasInternalFunctionality` and `hasUndocumented` in the `Parameters` class.
This means we now always have to output the parameters fields, even if previously skipped, so add an extra filter on the `data` field.

Change the `.info` field from `internalFunctionality` to `undocumented`