Anonymize configs before outputting them, either to console or file
Add .config.projectDir so we can always access it without always having to pass the path as an argument to functions
Set .config.projectDir to @internalFunctionality so it's never outputted (would defeat the whole purpose)
This could potentially clean up more code, but this is not done yet

What if there is no projectDir set (no viash project file found). We'll still be outputting full paths!